feat: add ability to kick member
